Improve automatic uploads and configuration files
The new implementation of automatic uploads is more flexible than
replacing substrings in the remote filename. With this implementation,
arbitrary scripts can be executed for each upload. This makes it
it possible to connect to different WebDAV servers based on the local
filename.

To implement this functionality, it was necessary to improve the
handling of configuration files. Configuration settings are now handled
in a class tdav.param that is instantiated as the click context object.